requirements.txt
numpy
scipy
pandas
scikit-learn
pyyaml
joblib
pytest
sphinx
